Refactor App component and remove unused screens and styles
- Updated App.tsx to utilize the new TabNavigator for improved navigation structure.
- Removed HomeScreen, ProfileScreen, SettingsScreen, and HomeCard components to streamline the codebase.
- Eliminated associated styles and store files to reduce clutter and enhance maintainability.
- Improved overall application organization by consolidating navigation logic.
