+++ App.tsx
@@ -1,59 +1,11 @@
-/* eslint-disable react/no-unstable-nested-components */
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import AntDesign from 'react-native-vector-icons/AntDesign';
-import FontAwesome5 from 'react-native-vector-icons/FontAwesome5';
-
-// Import our screen components
-import HomeScreen from './screens/HomeScreen';
-import ProfileScreen from './screens/ProfileScreen';
-import SettingsScreen from './screens/SettingsScreen';
-
-// Create the bottom tab navigator
-const Tab = createBottomTabNavigator();
+import { TabNavigator } from './src/navigation';
 
 function App() {
   return (
     <NavigationContainer>
-      <Tab.Navigator
-        screenOptions={() => ({
-          tabBarActiveTintColor: '#007bff',
-          tabBarInactiveTintColor: 'gray',
-          headerShown: false,
-        })}
-      >
-        <Tab.Screen
-          name="Home"
-          component={HomeScreen}
-          options={{
-            tabBarIcon: ({ color, size }) => {
-              let iconName: string = 'home';
-              return <AntDesign name={iconName} size={size} color={color} />;
-            },
-          }}
-        />
-        <Tab.Screen
-          name="History"
-          component={ProfileScreen}
-          options={{
-            tabBarIcon: ({ color, size }) => {
-              let iconName: string = 'history';
-              return <FontAwesome5 name={iconName} size={size} color={color} />;
-            },
-          }}
-        />
-        <Tab.Screen
-          name="Settings"
-          component={SettingsScreen}
-          options={{
-            tabBarIcon: ({ color, size }) => {
-              let iconName = 'setting';
-              return <AntDesign name={iconName} size={size} color={color} />;
-            },
-          }}
-        />
-      </Tab.Navigator>
+      <TabNavigator />
     </NavigationContainer>
   );
 }
