+++ src/components/index.ts
@@ -1,0 +1,5 @@
+// Cards
+export { default as HomeCard } from './cards/HomeCard';
+
+// UI Components
+// Add future UI components here
