+++ src/navigation/index.ts
@@ -1,0 +1,1 @@
+export { default as TabNavigator } from './TabNavigator';
