+++ src/screens/index.ts
@@ -1,0 +1,3 @@
+export { default as HomeScreen } from './HomeScreen';
+export { default as ProfileScreen } from './ProfileScreen';
+export { default as SettingsScreen } from './SettingsScreen';
