+++ src/store/index.ts
@@ -1,0 +1,12 @@
+// Export all types
+export * from './types/types';
+
+// Export storage service
+export { StorageService } from './storage';
+
+// Export transaction store
+export { TransactionStore } from './slices/transactionStore';
+
+// Export theme store
+export { ThemeStore } from './slices/themeStore';
+export type { ThemeType } from './slices/themeStore';
