+++ src/styles/common/inputs.ts
