+++ src/styles/common/layout.ts
