+++ src/styles/components/homeCard.ts
