+++ src/styles/components/homeScreen.ts
