+++ src/styles/components/index.ts
