+++ src/styles/components/settingsScreen.ts
