+++ src/types/index.ts
@@ -1,0 +1,4 @@
+// Store types
+export * from '../store/types/types';
+
+// Add other application types here
