--- store/index.ts
@@ -1,12 +1,0 @@
-// Export all types
-export * from './types';
-
-// Export storage service
-export { StorageService } from './storage';
-
-// Export transaction store
-export { TransactionStore } from './transactionStore';
-
-// Export theme store
-export { ThemeStore } from './themeStore';
-export type { ThemeType } from './themeStore';
