--- styles/components/index.ts
@@ -1,6 +1,0 @@
-export { HomeCardStyles } from './homeCard';
-
-// Combined component styles
-export const ComponentStyles = {
-  homeCard: HomeCardStyles,
-};
